feat: 하나의 컴포넌트에 많은 역할들을 하는 컴포넌트가 모여있어서 각 컴포넌트로 분리
