+++ src/app/social/createSocialModal/type.ts
@@ -1,0 +1,7 @@
+import { GenreType } from '@/api/social/type';
+import { UseFormRegisterReturn } from 'react-hook-form';
+
+export interface SelectGenreProps {
+  register: UseFormRegisterReturn<'location'>;
+  handleChangeGenre: (genre: GenreType) => void;
+}
